pyproject.toml
[project]
name = "my-db-handler"
version = "0.1.0"
description = "Közös adatbázis és tároló kezelő modul (PostgreSQL & MinIO)"
authors = [
    {name = "Zoltan Pinter", email = "pizoli82@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
license = "Apache-2.0"
dependencies = [
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "minio (>=7.2.12,<8.0.0)"
]

[tool.poetry]
packages = [{include = "db_handler", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
